refactor: 상수파일 분리 해제
